+++ sponge-8/src/main/java/dev/qixils/crowdcontrol/plugin/sponge8/commands/ClutterCommand.java
@@ -3,23 +3,20 @@ package dev.qixils.crowdcontrol.plugin.sponge8.commands;
 import dev.qixils.crowdcontrol.common.ExecuteUsing;
 import dev.qixils.crowdcontrol.plugin.sponge8.ImmediateCommand;
 import dev.qixils.crowdcontrol.plugin.sponge8.SpongeCrowdControlPlugin;
-import dev.qixils.crowdcontrol.plugin.sponge8.utils.ItemUtil;
 import dev.qixils.crowdcontrol.socket.Request;
 import dev.qixils.crowdcontrol.socket.Response;
 import lombok.Getter;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 import org.spongepowered.api.entity.living.player.server.ServerPlayer;
 import org.spongepowered.api.event.CauseStackManager.StackFrame;
+import org.spongepowered.api.item.inventory.Inventory;
 import org.spongepowered.api.item.inventory.ItemStack;
-import org.spongepowered.api.item.inventory.Slot;
-import org.spongepowered.api.item.inventory.entity.PrimaryPlayerInventory;
-import org.spongepowered.api.item.inventory.type.GridInventory;
-import org.spongepowered.math.vector.Vector2i;
+import org.spongepowered.api.item.inventory.entity.PlayerInventory;
 
-import java.util.*;
-
-import static dev.qixils.crowdcontrol.common.command.CommandConstants.CLUTTER_ITEMS;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Getter
 @ExecuteUsing(ExecuteUsing.Type.SYNC_GLOBAL)
@@ -30,51 +27,6 @@ public class ClutterCommand extends ImmediateCommand {
 		super(plugin);
 	}
 
-	private static Vector2i uniqueSlot(GridInventory inventory, Collection<Vector2i> usedSlots) {
-		int rows = inventory.rows();
-		int columns = inventory.columns();
-		List<Vector2i> allSlots = new ArrayList<>(rows * columns);
-		for (int x = 0; x < columns; x++) {
-			for (int y = 0; y < rows; y++) {
-				allSlots.add(Vector2i.from(x, y));
-			}
-		}
-		Collections.shuffle(allSlots, random);
-
-		for (Vector2i slot : allSlots) {
-			if (!usedSlots.contains(slot))
-				return slot;
-		}
-		throw new IllegalArgumentException("All slots have been used");
-	}
-
-	private static Slot unwrap(@Nullable Slot slot, Vector2i pos) {
-		if (slot == null)
-			throw new IndexOutOfBoundsException("Slot " + pos.x() + "," + pos.y() + " is out of bounds");
-		return slot;
-	}
-
-	private static Slot unwrap(GridInventory inventory, Vector2i pos) {
-		return unwrap(inventory.slot(pos).orElse(null), pos);
-	}
-
-	private static boolean swap(GridInventory inv, Vector2i slotPos1, Vector2i slotPos2) {
-		Slot slot1 = unwrap(inv, slotPos1);
-		Slot slot2 = unwrap(inv, slotPos2);
-
-		ItemStack slot1item = slot1.poll().polledItem().createStack();
-		ItemStack slot2item = slot2.poll().polledItem().createStack();
-
-		if (ItemUtil.isSimilar(slot1item, slot2item))
-			return false;
-
-		// this is throwing warnings in the console which I cannot silence >.>
-		slot2.offer(slot1item);
-		slot1.offer(slot2item);
-
-		return true;
-	}
-
 	@Override
 	public Response.@NotNull Builder executeImmediately(@NotNull List<@NotNull ServerPlayer> players, @NotNull Request request) {
 		// swaps random items in player's inventory
@@ -82,23 +34,17 @@ public class ClutterCommand extends ImmediateCommand {
 		try (StackFrame frame = plugin.getGame().server().causeStackManager().pushCauseFrame()) {
 			frame.pushCause(plugin.getPluginContainer());
 			for (ServerPlayer player : players) {
-				PrimaryPlayerInventory inventory = player.inventory().primary();
-				GridInventory gridInventory = inventory.asGrid();
-				Set<Vector2i> swappedSlots = new HashSet<>(CLUTTER_ITEMS);
+				PlayerInventory inventory = player.inventory();
+				List<ItemStack> original = inventory.slots().stream().map(Inventory::peek).collect(Collectors.toList());
+				List<ItemStack> shuffled = new ArrayList<>(original);
+				Collections.shuffle(shuffled);
 
-				Vector2i heldItemSlot = Vector2i.from(inventory.hotbar().selectedSlotIndex(), 3);
-				swappedSlots.add(heldItemSlot);
-				Vector2i swapItemWith = uniqueSlot(gridInventory, swappedSlots);
-				swappedSlots.add(swapItemWith);
-				success |= swap(gridInventory, heldItemSlot, swapItemWith);
+				if (shuffled.equals(original)) continue;
 
-				while (swappedSlots.size() < CLUTTER_ITEMS) {
-					Vector2i newSlot1 = uniqueSlot(gridInventory, swappedSlots);
-					swappedSlots.add(newSlot1);
-					Vector2i newSlot2 = uniqueSlot(gridInventory, swappedSlots);
-					swappedSlots.add(newSlot2);
-					success |= swap(gridInventory, newSlot1, newSlot2);
+				for (int i = 0; i < shuffled.size(); i++) {
+					inventory.set(i, shuffled.get(i));
 				}
+				success = true;
 			}
 			frame.popCause();
 		}