+++ sponge-8/src/main/java/dev/qixils/crowdcontrol/plugin/sponge8/commands/package-info.java
@@ -1,4 +1,0 @@
-@ParametersAreNonnullByDefault
-package dev.qixils.crowdcontrol.plugin.sponge8.commands;
-
-import javax.annotation.ParametersAreNonnullByDefault;