+++ CHANGELOG.md
@@ -5,6 +5,7 @@ A list of all changes made to the software in reverse chronological order.
 ## 3.4.3
 
 - Added support for 1.20.6 on Sponge(Forge) 🥳
+- Updated the Clutter Inventory effect to shuffle around all items instead of just a few
 - Fixed issue where taking max health would successfully activate but refund anyway
 
 ## 3.4.2